Fixed 64-bit compile issue. later truncation is still fine, just about a hash value.
svn path=/trunk/KDE/kdelibs/; revision=888129

+++ khtml/css/css_base.h
@@ -280,7 +280,7 @@ namespace DOM {
         {
             unsigned result = 0;
             while (selector) {
-                result ^= (unsigned)selector->value.impl();
+                result ^= (unsigned long)selector->value.impl();
                 result ^= (selector->attrLocalName.id() << 3);
                 result ^= (selector->attrNamespace.id() << 7);
                 result ^= (selector->tagLocalName.id() << 10);
